avoid add new dependency
